+++ src/v1.x/packages/shared/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @copilotkit/shared
 
+## 1.51.4-next.2
+
 ## 1.51.4-next.1
 
 ## 1.51.4-next.0