+++ src/v2.x/packages/web-inspector/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkitnext/web-inspector
 
+## 1.51.4-next.2
+
+### Patch Changes
+
+- @copilotkitnext/core@1.51.4-next.2
+
 ## 1.51.4-next.1
 
 ### Patch Changes